forensic_hawkeye_env/server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pymunk>=6.6.0
pydantic>=2.0.0
